refactor: move Maths function to Maths folder #90
