+++ src/Maths/factorial.js
