+++ src/Maths/index.js
@@ -1,0 +1,48 @@
+import { average } from "./average.js";
+import { clamp } from "./clamp.js";
+import { degreesToRadians } from "./degreesToRadians.js";
+import { divideEvenly } from "./divideEvenly.js";
+import { division } from "./division.js";
+import { getPositionWithAngleDistance } from "./getPositionWithAngleDistance.js";
+import { isEven, isOdd } from "./isEvenOdd.js";
+import { invertedLerp } from "./invertedLerp.js";
+import { isMultipleOf } from "./isMultipleOf.js";
+import { isNarcissisticNumber } from "./isNarcissisticNumber.js";
+import { lerp } from "./lerp.js";
+import { makeNegative } from "./makeNegative.js";
+import { map } from "./map.js";
+import { median } from "./median.js";
+import { mode } from "./mode.js";
+import { multiplication } from "./multiplication.js";
+import { radiansToDegrees } from "./radiansToDegrees.js";
+import { range } from "./range.js";
+import { subtraction } from "./subtraction.js";
+import { sum } from "./sum.js";
+
+/**
+ * Math functions
+ * 
+ * @namespace Maths
+ */
+export {
+	average,
+	clamp,
+	degreesToRadians,
+	divideEvenly,
+	division,
+	getPositionWithAngleDistance,
+	invertedLerp,
+	isEven, isOdd,
+	isMultipleOf,
+	isNarcissisticNumber,
+	lerp,
+	makeNegative,
+	map,
+	median,
+	mode,
+	multiplication,
+	radiansToDegrees,
+	range,
+	subtraction,
+	sum,
+};
