+++ src/Maths/makeNegative.js
@@ -8,5 +8,8 @@
  * 
  * @param {number} number - The number you want to convert
  * @returns {number} - a negative number
+ * 
+ * @function makeNegative
+ * @memberof Maths
  */
 export function makeNegative(number) { return Math.abs(number) * -1; };
