+++ src/index.js
@@ -1,5 +1,6 @@
 import * as DOM from "./DOM/index.js";
 import * as Games from "./Games/index.js";
+import * as Maths from "./Maths/index.js";
 import * as SortingAlgorithms from "./SortingAlgorithms/index.js";
 
 import { allCharactersSame } from "./allCharactersSame.js";
@@ -14,56 +15,37 @@ import { arraySortAscending } from "./arraySortAscending.js";
 import { arraySortAscendingObject } from "./arraySortAscendingObject.js";
 import { arraySortDescending } from "./arraySortDescending.js";
 import { arraySortDescendingObject } from "./arraySortDescendingObject.js";
-import { average } from "./average.js";
 import { binary2Decimal } from "./binary2Decimal.js";
 import { BinarySearchTree, BinarySearchTreeInstance } from "./BinarySearchTree.js";
 import { stringToCamelCase, camelCaseToNormal } from "./camelCase.js";
 import { stringToCapitalize, capitalizeToNormal } from "./capitalizeCase.js";
-import { clamp } from "./clamp.js";
 import { clone } from "./clone.js";
 import { compare2Objects } from "./compare2Objects.js";
 import { decimal2Binary } from "./decimal2Binary.js";
-import { degreesToRadians } from "./degreesToRadians.js";
-import { division } from "./division.js";
-import { divideEvenly } from "./divideEvenly.js";
 import { EventSystem, EventSystemInstance } from "./EventSystem.js";
 import { Fibonacci, fibonacciSequence, fibonacciUntil, fibonacciCustomSequence, recursiveFibonacci } from "./Fibonacci.js";
 import { FIFO } from "./FIFO.js";
 import { getDate } from "./getDate.js"; // import { getTime, getMilliseconds, getSeconds, getMinutes, getHours, getDay, getWeekDay, getWeek, getMonth, getYear, getDateFormatted } from "./getDate.js";
-import { getPositionWithAngleDistance } from "./getPositionWithAngleDistance.js";
 import { getUrlParameter } from "./getUrlParameter.js";
 import { getVersion } from "./getVersion.js";
-import { invertedLerp } from "./invertedLerp.js";
 import { invertSentence, invertWords } from "./invertText.js";
-import { isEven, isOdd } from "./isEvenOdd.js";
-import { isMultipleOf } from "./isMultipleOf.js";
-import { isNarcissisticNumber } from "./isNarcissisticNumber.js";
 import { isFalsy, isTruthy } from "./isTruthyFalsy.js";
 import { stringToKebabCase, kebabCaseToNormal } from "./kebabCase.js";
-import { lerp } from "./lerp.js";
 import { and, or, xor } from "./logicalOperators.js";
 import { LIFO } from "./LIFO.js";
-import { makeNegative } from "./makeNegative.js";
-import { map } from "./map.js";
-import { median } from "./median.js";
-import { mode } from "./mode.js";
-import { multiplication } from "./multiplication.js";
-import { radiansToDegrees } from "./radiansToDegrees.js";
 import { randomColor, randomColor0X, randomRGBColor, randomRGBAColor } from "./randomColor.js";
 import { randomNumber, randomInt, randomFloat } from "./randomNumber.js";
 import { randomString } from "./randomString.js";
 import { randomWalk1D, randomWalk2D, randomWalk3D } from "./randomWalk.js";
-import { range } from "./range.js";
 import { stringToScreamingSnakeCase, screamingSnakeCaseToNormal } from "./screamingSnakeCase.js";
 import { stringToSnakeCase, snakeCaseToNormal } from "./snakeCase.js";
-import { subtraction } from "./subtraction.js";
-import { sum } from "./sum.js";
 import { topDownCarMovimentation } from "./topDownCarMovimentation.js";
 import { Vector2 } from "./Vector2.js";
 
 export {
 	DOM,
 	Games,
+	Maths,
 	SortingAlgorithms,
 
 	BinarySearchTree, BinarySearchTreeInstance,
@@ -83,44 +65,24 @@ export {
 	arrayShuffle,
 	arraySortAscending, arraySortAscendingObject,
 	arraySortDescending, arraySortDescendingObject,
-	average,
 	binary2Decimal,
 	stringToCamelCase, camelCaseToNormal,
 	stringToCapitalize, capitalizeToNormal,
-	clamp,
 	clone,
 	compare2Objects,
 	decimal2Binary,
-	degreesToRadians,
-	division,
-	divideEvenly,
 	getDate, // getTime, getMilliseconds, getSeconds, getMinutes, getHours, getDay, getWeekDay, getWeek, getMonth, getYear, getDateFormatted,
-	getPositionWithAngleDistance,
 	getUrlParameter,
 	getVersion,
-	invertedLerp,
 	invertSentence, invertWords,
-	isEven, isOdd,
-	isMultipleOf,
-	isNarcissisticNumber,
 	isFalsy, isTruthy,
 	stringToKebabCase, kebabCaseToNormal,
-	lerp,
 	and, or, xor, // logicalOperators
-	makeNegative,
-	map,
-	median,
-	mode,
-	multiplication,
-	radiansToDegrees,
 	randomColor, randomColor0X, randomRGBColor, randomRGBAColor,
 	randomNumber, randomInt, randomFloat,
 	randomString,
 	randomWalk1D, randomWalk2D, randomWalk3D,
-	range,
 	stringToScreamingSnakeCase, screamingSnakeCaseToNormal,
 	stringToSnakeCase, snakeCaseToNormal,
-	subtraction,
-	sum,
 	topDownCarMovimentation,
 };
